+++ realm/realm-library/src/androidTest/java/io/realm/RealmConfigurationTests.java
@@ -543,7 +543,7 @@ public class RealmConfigurationTests {
     public void modelClassesForDefaultMediator() throws Exception {
         assertTrue(defaultConfig.getSchemaMediator() instanceof DefaultRealmModuleMediator);
 
-        final Set<Class<? extends RealmObject>> realmClasses = defaultConfig.getRealmObjectClasses();
+        final Set<Class<? extends RealmModel>> realmClasses = defaultConfig.getRealmObjectClasses();
 
         assertTrue(realmClasses.contains(AllTypes.class));
 
@@ -561,7 +561,7 @@ public class RealmConfigurationTests {
                 .setModules(new HumanModule()).build();
         assertTrue(config.getSchemaMediator() instanceof HumanModuleMediator);
 
-        final Set<Class<? extends RealmObject>> realmClasses = config.getRealmObjectClasses();
+        final Set<Class<? extends RealmModel>> realmClasses = config.getRealmObjectClasses();
 
         assertFalse(realmClasses.contains(AllTypes.class));
         assertTrue(realmClasses.contains(CatOwner.class));
@@ -581,7 +581,7 @@ public class RealmConfigurationTests {
                 .setModules(new HumanModule(), new AnimalModule()).build();
         assertTrue(config.getSchemaMediator() instanceof CompositeMediator);
 
-        final Set<Class<? extends RealmObject>> realmClasses = config.getRealmObjectClasses();
+        final Set<Class<? extends RealmModel>> realmClasses = config.getRealmObjectClasses();
 
         assertFalse(realmClasses.contains(AllTypes.class));
         assertTrue(realmClasses.contains(CatOwner.class));
@@ -602,7 +602,7 @@ public class RealmConfigurationTests {
                 .schema(AllTypes.class, CatOwner.class).build();
         assertTrue(config.getSchemaMediator() instanceof FilterableMediator);
 
-        final Set<Class<? extends RealmObject>> realmClasses = config.getRealmObjectClasses();
+        final Set<Class<? extends RealmModel>> realmClasses = config.getRealmObjectClasses();
 
         assertTrue(realmClasses.contains(AllTypes.class));
         assertTrue(realmClasses.contains(CatOwner.class));
@@ -630,7 +630,7 @@ public class RealmConfigurationTests {
             }
 
             @Override
-            public <E extends RealmObject> Observable<RealmResults<E>> from(Realm realm, RealmResults<E> results) {
+            public <E extends RealmModel> Observable<RealmResults<E>> from(Realm realm, RealmResults<E> results) {
                 return null;
             }
 
@@ -640,7 +640,7 @@ public class RealmConfigurationTests {
             }
 
             @Override
-            public <E extends RealmObject> Observable<RealmList<E>> from(Realm realm, RealmList<E> list) {
+            public <E extends RealmModel> Observable<RealmList<E>> from(Realm realm, RealmList<E> list) {
                 return null;
             }
 
@@ -650,7 +650,7 @@ public class RealmConfigurationTests {
             }
 
             @Override
-            public <E extends RealmObject> Observable<E> from(Realm realm, E object) {
+            public <E extends RealmModel> Observable<E> from(Realm realm, E object) {
                 return null;
             }
 
@@ -660,7 +660,7 @@ public class RealmConfigurationTests {
             }
 
             @Override
-            public <E extends RealmObject> Observable<RealmQuery<E>> from(Realm realm, RealmQuery<E> query) {
+            public <E extends RealmModel> Observable<RealmQuery<E>> from(Realm realm, RealmQuery<E> query) {
                 return null;
             }