+++ realm/realm-library/src/androidTest/java/io/realm/RealmObjectTests.java
@@ -46,6 +46,7 @@ import io.realm.entities.Dog;
 import io.realm.entities.NullTypes;
 import io.realm.entities.StringAndInt;
 import io.realm.entities.Thread;
+import io.realm.internal.RealmObjectProxy;
 import io.realm.internal.Row;
 import io.realm.internal.Table;
 import io.realm.rule.RunInLooperThread;
@@ -102,8 +103,8 @@ public class RealmObjectTests {
     @Test
     public void row_isValid() {
         realm.beginTransaction();
-        RealmObject realmObject = realm.createObject(AllTypes.class);
-        Row row = realmObject.row;
+        RealmObjectProxy realmObject = (RealmObjectProxy) realm.createObject(AllTypes.class);
+        Row row = realmObject.realmGet$proxyState().getRow$realm();
         realm.commitTransaction();
 
         assertNotNull("RealmObject.realmGetRow returns zero ", row);
@@ -225,7 +226,7 @@ public class RealmObjectTests {
         try {
             dog.deleteFromRealm();
             fail("Failed on deleting a RealmObject from null Row.");
-        } catch (IllegalStateException ignored) {
+        } catch (IllegalArgumentException ignored) {
         }
     }