+++ realm/realm-library/src/androidTest/java/io/realm/TypeBasedNotificationsTests.java
@@ -43,6 +43,7 @@ import io.realm.entities.Cat;
 import io.realm.entities.Dog;
 import io.realm.entities.Owner;
 import io.realm.entities.PrimaryKeyAsLong;
+import io.realm.internal.RealmObjectProxy;
 import io.realm.proxy.HandlerProxy;
 import io.realm.rule.RunInLooperThread;
 import io.realm.rule.RunTestInLooperThread;
@@ -1698,7 +1699,7 @@ public class TypeBasedNotificationsTests {
                 Cat cat = owner.getCat();
                 boolean foundKey = false;
                 // Check if cat has been added to the realmObjects in case of the behaviour of getCat changes
-                for (WeakReference<RealmObject> weakReference : realm.handlerController.realmObjects.keySet()) {
+                for (WeakReference<RealmObjectProxy> weakReference : realm.handlerController.realmObjects.keySet()) {
                     if (weakReference.get() == cat) {
                         foundKey = true;
                         break;
@@ -1735,7 +1736,7 @@ public class TypeBasedNotificationsTests {
                 boolean foundKey = false;
                 // Check if the results has been added to the syncRealmResults in case of the behaviour of
                 // allObjects changes
-                for (WeakReference<RealmResults<? extends RealmObject>> weakReference :
+                for (WeakReference<RealmResults<? extends RealmModel>> weakReference :
                         realm.handlerController.syncRealmResults.keySet()) {
                     if (weakReference.get() == results) {
                         foundKey = true;