+++ realm/realm-library/src/main/java/io/realm/RealmList.java
@@ -25,6 +25,7 @@ import java.util.List;
 
 import io.realm.internal.InvalidRow;
 import io.realm.internal.LinkView;
+import io.realm.internal.RealmObjectProxy;
 
 /**
  * RealmList is used to model one-to-many relationships in a {@link io.realm.RealmObject}.
@@ -44,7 +45,7 @@ import io.realm.internal.LinkView;
  * @param <E> the class of objects in list.
  */
 
-public class RealmList<E extends RealmObject> extends AbstractList<E> implements OrderedRealmCollection<E> {
+public final class RealmList<E extends RealmModel> extends AbstractList<E> implements OrderedRealmCollection<E> {
 
     private static final String ONLY_IN_MANAGED_MODE_MESSAGE = "This method is only available in managed mode";
     private static final String NULL_OBJECTS_NOT_ALLOWED_MESSAGE = "RealmList does not accept null values";
@@ -133,11 +134,11 @@ public class RealmList<E extends RealmObject> extends AbstractList<E> implements
      * <ol>
      * <li><b>Un-managed RealmLists:</b> It is possible to add both managed and un-managed objects. If adding managed
      * objects to a un-managed RealmList they will not be copied to the Realm again if using
-     * {@link Realm#copyToRealm(RealmObject)} afterwards.</li>
+     * {@link Realm#copyToRealm(RealmModel)} afterwards.</li>
      *
      * <li><b>Managed RealmLists:</b> It is possible to add un-managed objects to a RealmList that is already managed. In
-     * that case the object will transparently be copied to Realm using {@link Realm#copyToRealm(RealmObject)}
-     * or {@link Realm#copyToRealmOrUpdate(RealmObject)} if it has a primary key.</li>
+     * that case the object will transparently be copied to Realm using {@link Realm#copyToRealm(RealmModel)}
+     * or {@link Realm#copyToRealmOrUpdate(RealmModel)} if it has a primary key.</li>
      * </ol>
      *
      * @param location the index at which to insert.
@@ -153,8 +154,8 @@ public class RealmList<E extends RealmObject> extends AbstractList<E> implements
             if (location < 0 || location > size()) {
                 throw new IndexOutOfBoundsException("Invalid index " + location + ", size is " + size());
             }
-            object = copyToRealmIfNeeded(object);
-            view.insert(location, object.row.getIndex());
+            RealmObjectProxy proxy = (RealmObjectProxy) copyToRealmIfNeeded(object);
+            view.insert(location, proxy.realmGet$proxyState().getRow$realm().getIndex());
         } else {
             nonManagedList.add(location, object);
         }
@@ -165,11 +166,11 @@ public class RealmList<E extends RealmObject> extends AbstractList<E> implements
      * <ol>
      * <li><b>Un-managed RealmLists:</b> It is possible to add both managed and un-managed objects. If adding managed
      * objects to a un-managed RealmList they will not be copied to the Realm again if using
-     * {@link Realm#copyToRealm(RealmObject)} afterwards.</li>
+     * {@link Realm#copyToRealm(RealmModel)} afterwards.</li>
      *
      * <li><b>Managed RealmLists:</b> It is possible to add un-managed objects to a RealmList that is already managed. In
-     * that case the object will transparently be copied to Realm using {@link Realm#copyToRealm(RealmObject)}
-     * or {@link Realm#copyToRealmOrUpdate(RealmObject)} if it has a primary key.</li>
+     * that case the object will transparently be copied to Realm using {@link Realm#copyToRealm(RealmModel)}
+     * or {@link Realm#copyToRealmOrUpdate(RealmModel)} if it has a primary key.</li>
      * </ol>
      *
      * @param object the object to add.
@@ -181,8 +182,8 @@ public class RealmList<E extends RealmObject> extends AbstractList<E> implements
         checkValidObject(object);
         if (managedMode) {
             checkValidView();
-            object = copyToRealmIfNeeded(object);
-            view.add(object.row.getIndex());
+            RealmObjectProxy proxy = (RealmObjectProxy) copyToRealmIfNeeded(object);
+            view.add(proxy.realmGet$proxyState().getRow$realm().getIndex());
         } else {
             nonManagedList.add(object);
         }
@@ -194,11 +195,11 @@ public class RealmList<E extends RealmObject> extends AbstractList<E> implements
      * <ol>
      * <li><b>Un-managed RealmLists:</b> It is possible to add both managed and un-managed objects. If adding managed
      * objects to a un-managed RealmList they will not be copied to the Realm again if using
-     * {@link Realm#copyToRealm(RealmObject)} afterwards.</li>
+     * {@link Realm#copyToRealm(RealmModel)} afterwards.</li>
      *
      * <li><b>Managed RealmLists:</b> It is possible to add un-managed objects to a RealmList that is already managed.
-     * In that case the object will transparently be copied to Realm using {@link Realm#copyToRealm(RealmObject)} or
-     * {@link Realm#copyToRealmOrUpdate(RealmObject)} if it has a primary key.</li>
+     * In that case the object will transparently be copied to Realm using {@link Realm#copyToRealm(RealmModel)} or
+     * {@link Realm#copyToRealmOrUpdate(RealmModel)} if it has a primary key.</li>
      * </ol>
      * @param location the index at which to put the specified object.
      * @param object the object to add.
@@ -211,9 +212,9 @@ public class RealmList<E extends RealmObject> extends AbstractList<E> implements
         checkValidObject(object);
         if (managedMode) {
             checkValidView();
-            object = copyToRealmIfNeeded(object);
+            RealmObjectProxy proxy = (RealmObjectProxy) copyToRealmIfNeeded(object);
             E oldObject = get(location);
-            view.set(location, object.row.getIndex());
+            view.set(location, proxy.realmGet$proxyState().getRow$realm().getIndex());
             return oldObject;
         } else {
             return nonManagedList.set(location, object);
@@ -222,33 +223,37 @@ public class RealmList<E extends RealmObject> extends AbstractList<E> implements
 
     // Transparently copies a standalone object or managed object from another Realm to the Realm backing this RealmList.
     private E copyToRealmIfNeeded(E object) {
-        if (object instanceof DynamicRealmObject) {
-            String listClassName = RealmSchema.getSchemaForTable(view.getTargetTable());
-            String objectClassName = ((DynamicRealmObject) object).getType();
-            if (object.realm == realm) {
-                if (listClassName.equals(objectClassName)) {
-                    // Same Realm instance and same target table
-                    return object;
+        if (object instanceof RealmObjectProxy) {
+            RealmObjectProxy proxy = (RealmObjectProxy) object;
+
+            if (proxy instanceof DynamicRealmObject) {
+                String listClassName = RealmSchema.getSchemaForTable(view.getTargetTable());
+                String objectClassName = ((DynamicRealmObject) object).getType();
+                if (proxy.realmGet$proxyState().getRealm$realm() == realm) {
+                    if (listClassName.equals(objectClassName)) {
+                        // Same Realm instance and same target table
+                        return object;
+                    } else {
+                        // Different target table
+                        throw new IllegalArgumentException(String.format("The object has a different type from list's." +
+                                " Type of the list is '%s', type of object is '%s'.", listClassName, objectClassName));
+                    }
+                } else if (realm.threadId == proxy.realmGet$proxyState().getRealm$realm().threadId) {
+                    // We don't support moving DynamicRealmObjects across Realms automatically. The overhead is too big as
+                    // you have to run a full schema validation for each object.
+                    // And copying from another Realm instance pointed to the same Realm file is not supported as well.
+                    throw new IllegalArgumentException("Cannot copy DynamicRealmObject between Realm instances.");
                 } else {
-                    // Different target table
-                    throw new IllegalArgumentException(String.format("The object has a different type from list's." +
-                            " Type of the list is '%s', type of object is '%s'.", listClassName, objectClassName));
+                    throw new IllegalStateException("Cannot copy an object to a Realm instance created in another thread.");
                 }
-            } else if (realm.threadId == object.realm.threadId) {
-                // We don't support moving DynamicRealmObjects across Realms automatically. The overhead is too big as
-                // you have to run a full schema validation for each object.
-                // And copying from another Realm instance pointed to the same Realm file is not supported as well.
-                throw new IllegalArgumentException("Cannot copy DynamicRealmObject between Realm instances.");
             } else {
-                throw new IllegalStateException("Cannot copy an object to a Realm instance created in another thread.");
-            }
-        } else {
-            // Object is already in this realm
-            if (object.row != null && object.realm.getPath().equals(realm.getPath())) {
-                if (realm != object.realm) {
-                    throw new IllegalArgumentException("Cannot copy an object from another Realm instance.");
+                // Object is already in this realm
+                if (proxy.realmGet$proxyState().getRow$realm() != null && proxy.realmGet$proxyState().getRealm$realm().getPath().equals(realm.getPath())) {
+                    if (realm != proxy.realmGet$proxyState().getRealm$realm()) {
+                        throw new IllegalArgumentException("Cannot copy an object from another Realm instance.");
+                    }
+                    return object;
                 }
-                return object;
             }
         }
 
@@ -670,10 +675,10 @@ public class RealmList<E extends RealmObject> extends AbstractList<E> implements
         boolean contains = false;
         if (managedMode) {
             realm.checkIfValid();
-            if (object instanceof RealmObject) {
-                RealmObject realmObject = (RealmObject) object;
-                if (realmObject.row != null && realm.getPath().equals(realmObject.realm.getPath()) && realmObject.row != InvalidRow.INSTANCE) {
-                    contains = view.contains(realmObject.row.getIndex());
+            if (object instanceof RealmObjectProxy) {
+                RealmObjectProxy proxy = (RealmObjectProxy) object;
+                if (proxy.realmGet$proxyState().getRow$realm() != null && realm.getPath().equals(proxy.realmGet$proxyState().getRealm$realm().getPath()) && proxy.realmGet$proxyState().getRow$realm() != InvalidRow.INSTANCE) {
+                    contains = view.contains(proxy.realmGet$proxyState().getRow$realm().getIndex());
                 }
             }
         } else {
@@ -712,7 +717,7 @@ public class RealmList<E extends RealmObject> extends AbstractList<E> implements
         } else {
             for (int i = 0; i < size(); i++) {
                 if (managedMode) {
-                    sb.append(get(i).row.getIndex());
+                    sb.append(((RealmObjectProxy) get(i)).realmGet$proxyState().getRow$realm().getIndex());
                 } else {
                     sb.append(System.identityHashCode(get(i)));
                 }